@@ -9,75 +9,87 @@
  * file that was distributed with this source code.
  */
 
-namespace ICanBoogie;
+namespace Tests\ICanBoogie;
 
-class HelpersTest extends \PHPUnit_Framework_TestCase
+use PHPUnit\Framework\TestCase;
+
+use function ICanBoogie\array_flatten;
+use function ICanBoogie\escape;
+use function ICanBoogie\format;
+use function ICanBoogie\remove_accents;
+use function ICanBoogie\sort_by_weight;
+
+/**
+ * @group unit
+ */
+final class HelpersTest extends TestCase
 {
-	public function arrayProvider()
+	public function test_escape()
 	{
-		return array
-		(
-			array
-			(
-				array
-				(
-					'one' => array
-					(
-						'one' => 1,
-						'two' => array
-						(
+		$this->assertSame("&amp;&quot;'&lt;ab&gt;cd", escape('&"\'<ab>cd'));
+	}
+
+	public function array_provider(): array
+	{
+		return [
+			[
+				[
+					'one' =>
+						[
+							'one' => 1,
+							'two' =>
+								[
+									'one' => 1,
+									'two' => 2,
+									'three' =>
+										[
+											'one' => 1,
+											'two' => 2,
+											'three' => 3,
+										],
+								],
+
+							'three' =>
+								[
+									'one' =>
+										[
+											'one' => 1,
+											'two' => 2,
+											'three' => 3,
+										],
+
+									'two' =>
+										[
+											'one' => 1,
+											'two' => 2,
+											'three' => 3,
+										],
+
+									'three' => 3,
+								],
+						],
+
+					'two' =>
+						[
 							'one' => 1,
 							'two' => 2,
-							'three' => array
-							(
-								'one' => 1,
-								'two' => 2,
-								'three' => 3
-							)
-						),
-
-						'three' => array
-						(
-							'one' => array
-							(
-								'one' => 1,
-								'two' => 2,
-								'three' => 3
-							),
-
-							'two' => array
-							(
-								'one' => 1,
-								'two' => 2,
-								'three' => 3
-							),
-
-							'three' => 3
-						)
-					),
-
-					'two' => array
-					(
-						'one' => 1,
-						'two' => 2,
-						'three' => 3
-					)
-				)
-			)
-		);
+							'three' => 3,
+						],
+				],
+			],
+		];
 	}
 
 	/**
-	 * @dataProvider arrayProvider()
+	 * @dataProvider array_provider()
 	 */
-	public function testArrayFlatten($data)
+	public function test_array_flatten(array $data): void
 	{
 		$flat = array_flatten($data);
 
 		$this->assertEquals
 		(
-			array
-			(
+			[
 				'one.one' => 1,
 				'one.two.one' => 1,
 				'one.two.two' => 2,
@@ -93,24 +105,23 @@ class HelpersTest extends \PHPUnit_Framework_TestCase
 				'one.three.three' => 3,
 				'two.one' => 1,
 				'two.two' => 2,
-				'two.three' => 3
-			),
+				'two.three' => 3,
+			],
 
 			$flat
 		);
 	}
 
 	/**
-	 * @dataProvider arrayProvider()
+	 * @dataProvider array_provider()
 	 */
-	public function testArrayFlattenDouble($data)
+	public function test_array_flatten_double(array $data): void
 	{
-		$flat = array_flatten($data, array('[', ']'));
+		$flat = array_flatten($data, [ '[', ']' ]);
 
 		$this->assertEquals
 		(
-			array
-			(
+			[
 				'one[one]' => 1,
 				'one[two][one]' => 1,
 				'one[two][two]' => 2,
@@ -126,8 +137,8 @@ class HelpersTest extends \PHPUnit_Framework_TestCase
 				'one[three][three]' => 3,
 				'two[one]' => 1,
 				'two[two]' => 2,
-				'two[three]' => 3
-			),
+				'two[three]' => 3,
+			],
 
 			$flat
 		);
@@ -136,61 +147,54 @@ class HelpersTest extends \PHPUnit_Framework_TestCase
 	/**
 	 * @dataProvider provide_test_sort_by_weight
 	 */
-	public function test_sort_by_weight($array, $expected)
+	public function test_sort_by_weight(array $array, array $expected): void
 	{
-		$this->assertSame($expected, sort_by_weight($array, function($v) { return $v; }));
+		$this->assertSame($expected, sort_by_weight($array, function ($v) {
+			return $v;
+		}));
 	}
 
-	public function provide_test_sort_by_weight()
+	public function provide_test_sort_by_weight(): array
 	{
-		return array
-		(
+		return [
 			#1
 
-			array
-			(
-				array
-				(
+			[
+				[
 					'bottom' => 'bottom',
 					'min' => -10000,
 					'max' => 10000,
-					'top' => 'top'
-				),
+					'top' => 'top',
+				],
 
-				array
-				(
+				[
 					'top' => 'top',
 					'min' => -10000,
 					'max' => 10000,
-					'bottom' => 'bottom'
-				)
-			),
+					'bottom' => 'bottom',
+				],
+			],
 
 			#3: missing relative
 
-			array
-			(
-				array
-				(
+			[
+				[
 					'two' => 2,
 					'one' => 1,
-					'four' => 'after:three'
-				),
+					'four' => 'after:three',
+				],
 
-				array
-				(
+				[
 					'four' => 'after:three',
 					'one' => 1,
-					'two' => 2
-				)
-			),
+					'two' => 2,
+				],
+			],
 
 			#2
 
-			array
-			(
-				array
-				(
+			[
+				[
 					'two' => 0,
 					'three' => 0,
 					'bottom' => 'bottom',
@@ -200,11 +204,10 @@ class HelpersTest extends \PHPUnit_Framework_TestCase
 					'four' => 'after:three',
 					'top' => 'top',
 					'megatop' => 'top',
-					'hypertop' => 'top'
-				),
+					'hypertop' => 'top',
+				],
 
-				array
-				(
+				[
 					'hypertop' => 'top',
 					'megatop' => 'top',
 					'top' => 'top',
@@ -214,71 +217,65 @@ class HelpersTest extends \PHPUnit_Framework_TestCase
 					'four' => 'after:three',
 					'bottom' => 'bottom',
 					'megabottom' => 'bottom',
-					'hyperbottom' => 'bottom'
-				)
-			)
-		);
+					'hyperbottom' => 'bottom',
+				],
+			],
+		];
 	}
 
 	/**
 	 * @dataProvider provide_test_remove_accents
 	 */
-	public function test_remove_accents($expected, $str)
+	public function test_remove_accents(string $expected, string $str): void
 	{
 		$this->assertEquals($expected, remove_accents($str));
 	}
 
-	public function provide_test_remove_accents()
+	public function provide_test_remove_accents(): array
 	{
-		return array
-		(
-			array('AAAAAAAE', 'ÁÀÂÄÃÅÆ'),
-			array('aaaaaaae', 'áàâäãåæ'),
-			array('C', 'Ç'),
-			array('c', 'ç'),
-			array('EEEE', 'ÉÈÊË'),
-			array('eeee', 'éèêë'),
-			array('IIII', 'ÍÏÎÌ'),
-			array('iiii', 'íìîï'),
-			array('N', 'Ñ'),
-			array('n', 'ñ'),
-			array('OOOOO', 'ÓÒÔÖÕ'),
-			array('oooooo', 'óòôöõø'),
-			array('S', 'Š'),
-			array('s', 'š'),
-			array('UUUU', 'ÚÙÛÜ'),
-			array('uuuu', 'úùûü'),
-			array('YY', 'ÝŸ'),
-			array('yy', 'ýÿ')
-		);
+		return [
+
+			[ 'AAAAAAAE', 'ÁÀÂÄÃÅÆ' ],
+			[ 'aaaaaaae', 'áàâäãåæ' ],
+			[ 'C', 'Ç' ],
+			[ 'c', 'ç' ],
+			[ 'EEEE', 'ÉÈÊË' ],
+			[ 'eeee', 'éèêë' ],
+			[ 'IIII', 'ÍÏÎÌ' ],
+			[ 'iiii', 'íìîï' ],
+			[ 'N', 'Ñ' ],
+			[ 'n', 'ñ' ],
+			[ 'OOOOO', 'ÓÒÔÖÕ' ],
+			[ 'oooooo', 'óòôöõø' ],
+			[ 'S', 'Š' ],
+			[ 's', 'š' ],
+			[ 'UUUU', 'ÚÙÛÜ' ],
+			[ 'uuuu', 'úùûü' ],
+			[ 'YY', 'ÝŸ' ],
+			[ 'yy', 'ýÿ' ],
+
+		];
 	}
 
 	/**
 	 * @dataProvider provide_test_format
-	 *
-	 * @param string $format
-	 * @param array $args
-	 * @param string $expected
 	 */
-	public function test_format($format, array $args, $expected)
+	public function test_format(string $format, array $args, string $expected): void
 	{
 		$this->assertEquals($expected, format($format, $args));
 	}
 
-	/**
-	 * @return array
-	 */
-	public function provide_test_format()
+	public function provide_test_format(): array
 	{
-		return array(
+		return [
 
-			array("abc \\0", array( '<def' ), "abc <def"),
-			array("abc {0}", array( '<def' ), "abc <def"),
-			array("abc {a1}", array( 'a1' => '<def' ), "abc &lt;def"),
-			array("abc :a1", array( 'a1' => '<def' ), "abc <def"),
-			array("abc !a1", array( 'a1' => '<def' ), "abc &lt;def"),
-			array("abc %a1", array( 'a1' => '<def' ), "abc `&lt;def`")
+			[ "abc \\0", [ '<def' ], "abc <def" ],
+			[ "abc {0}", [ '<def' ], "abc <def" ],
+			[ "abc {a1}", [ 'a1' => '<def' ], "abc &lt;def" ],
+			[ "abc :a1", [ 'a1' => '<def' ], "abc <def" ],
+			[ "abc !a1", [ 'a1' => '<def' ], "abc &lt;def" ],
+			[ "abc %a1", [ 'a1' => '<def' ], "abc `&lt;def`" ],
 
-		);
+		];
 	}
 }
