@@ -11,13 +11,15 @@
 
 namespace ICanBoogie;
 
+use RuntimeException;
+
 /**
  * Exception thrown when there is something wrong with an array offset.
  *
  * This is the base class for offset exceptions, one should rather use the
  * {@link OffsetNotReadable} or {@link OffsetNotWritable} exceptions.
  */
-class OffsetError extends \RuntimeException
+class OffsetError extends RuntimeException
 {
 
 }
