@@ -11,6 +11,8 @@
 
 namespace ICanBoogie;
 
+use Throwable;
+
 /**
  * Exception thrown when a property is not writable.
  *
@@ -18,35 +20,27 @@ namespace ICanBoogie;
  */
 class PropertyNotWritable extends PropertyError
 {
-	public function __construct($message, $code=500, \Exception $previous=null)
+	public function __construct(string|array $message, Throwable $previous = null)
 	{
 		if (is_array($message))
 		{
-			list($property, $container) = $message + array(1 => null);
+			[ $property, $container ] = $message + [ 1 => null ];
 
 			if (is_object($container))
 			{
-				$message = format
-				(
-					'The property %property for object of class %class is not writable.', array
-					(
-						'%property' => $property,
-						'%class' => get_class($container)
-					)
-				);
+				$message = format("The property %property for object of class %class is not writable.", [
+					'%property' => $property,
+					'%class' => get_class($container)
+				]);
 			}
 			else
 			{
-				$message = format
-				(
-					'The property %property is not writable.', array
-					(
-						'%property' => $property
-					)
-				);
+				$message = format('The property %property is not writable.',[
+					'%property' => $property
+				]);
 			}
 		}
 
-		parent::__construct($message, $code, $previous);
+		parent::__construct($message, 0, $previous);
 	}
 }
